requirements.txt
flask
requests
openai-whisper
selenium
